+++ code/functions2.h
@@ -1,6 +1,6 @@
 #ifndef __FUNCTIONS2__
 #define __FUNCTIONS2__
 
-int find_hole(char* filename, int fd, size_t my_size);
+off_t find_hole(char* filename, int fd, size_t my_size);
 
 #endif
